Refactored upload clients to be filetype-agnostic
