+++ .gitignore
@@ -5,3 +5,4 @@ launch.json
 *.pyc
 __pycache__
 .pytest_cache
+data
