+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "dpytools"
-version = "0.4.5"
+version = "0.5.0"
 description = "Simple reusable python resources for digital publishing"
 authors = ["Your Name <you@example.com>"]
 license = "MIT"
