+++ tests/http/dataset/test_dataset_api.py
@@ -1,6 +1,5 @@
 import os
-from pathlib import Path
-from unittest.mock import MagicMock, mock_open, patch
+from unittest.mock import MagicMock, patch
 
 import pytest
 from requests import Response
