+++ tests/http/upload/test_upload_service_client.py
@@ -2,7 +2,7 @@ import os
 from pathlib import Path
 from unittest.mock import MagicMock, patch
 
-from dpytools.http.upload.base_upload import BaseUploadClient
+from dpytools.http.upload.upload_service_client import UploadServiceClient
 
 
 def mock_successful_token_response(*args, **kwargs):
@@ -19,24 +19,12 @@ def mock_successful_token_response(*args, **kwargs):
     return mock_response
 
 
-@patch("requests.request", side_effect=mock_successful_token_response)
-def test_base_upload_client_init(mock_request):
-    """
-    Ensures that the BaseUploadClient initializes correctly with the given parameters.
-    """
-    os.environ["FLORENCE_USER"] = "test_user"
-    os.environ["FLORENCE_PASSWORD"] = "test_password"
-    os.environ["IDENTITY_API_URL"] = "http://test_url"
-
-    client = BaseUploadClient(upload_url="http://example.com/upload", backoff_max=30)
-    assert client.upload_url == "http://example.com/upload"
-    assert client.token_auth.backoff_max == 30
-
-
-@patch("dpytools.http.upload.base_upload._create_temp_chunks")
-@patch("dpytools.http.upload.base_upload._delete_temp_chunks")
-@patch("dpytools.http.upload.base_upload._generate_upload_params")
-@patch("dpytools.http.upload.base_upload.BaseUploadClient._upload_file_chunks")
+@patch("dpytools.http.upload.upload_service_client._create_temp_chunks")
+@patch("dpytools.http.upload.upload_service_client._delete_temp_chunks")
+@patch("dpytools.http.upload.upload_service_client._generate_upload_params")
+@patch(
+    "dpytools.http.upload.upload_service_client.UploadServiceClient._upload_file_chunks"
+)
 @patch("requests.request", side_effect=mock_successful_token_response)
 def test_upload(
     mock_request,
@@ -52,17 +40,17 @@ def test_upload(
     os.environ["FLORENCE_PASSWORD"] = "test_password"
     os.environ["IDENTITY_API_URL"] = "http://test_url"
 
-    client = BaseUploadClient(upload_url="http://example.com/upload", backoff_max=30)
+    client = UploadServiceClient(upload_url="http://example.com/upload", backoff_max=30)
     mock_create_temp_chunks.return_value = ["chunk1", "chunk2"]
     mock_generate_upload_params.return_value = {"resumableIdentifier": "test_id"}
 
-    client._upload(file_path="test_file.csv", mimetype="text/csv")
+    client.upload(file_path="tests/test_cases/countries.csv", mimetype="text/csv")
 
     mock_create_temp_chunks.assert_called_once_with(
-        Path("test_file.csv").absolute(), 5242880
+        Path("tests/test_cases/countries.csv").absolute(), 5242880
     )
     mock_generate_upload_params.assert_called_once_with(
-        Path("test_file.csv").absolute(), "text/csv", 5242880
+        Path("tests/test_cases/countries.csv").absolute(), "text/csv", 5242880
     )
     mock_upload_file_chunks.assert_called_once_with(
         ["chunk1", "chunk2"], {"resumableIdentifier": "test_id"}
@@ -70,10 +58,12 @@ def test_upload(
     mock_delete_temp_chunks.assert_called_once_with(["chunk1", "chunk2"])
 
 
-@patch("dpytools.http.upload.base_upload._create_temp_chunks")
-@patch("dpytools.http.upload.base_upload._delete_temp_chunks")
-@patch("dpytools.http.upload.base_upload._generate_upload_new_params")
-@patch("dpytools.http.upload.base_upload.BaseUploadClient._upload_file_chunks")
+@patch("dpytools.http.upload.upload_service_client._create_temp_chunks")
+@patch("dpytools.http.upload.upload_service_client._delete_temp_chunks")
+@patch("dpytools.http.upload.upload_service_client._generate_upload_new_params")
+@patch(
+    "dpytools.http.upload.upload_service_client.UploadServiceClient._upload_file_chunks"
+)
 @patch("requests.request", side_effect=mock_successful_token_response)
 def test_upload_new(
     mock_request,
@@ -89,12 +79,12 @@ def test_upload_new(
     os.environ["FLORENCE_PASSWORD"] = "test_password"
     os.environ["IDENTITY_API_URL"] = "http://test_url"
 
-    client = BaseUploadClient(upload_url="http://example.com/upload", backoff_max=30)
+    client = UploadServiceClient(upload_url="http://example.com/upload", backoff_max=30)
     mock_create_temp_chunks.return_value = ["chunk1", "chunk2"]
     mock_generate_upload_new_params.return_value = {"Path": "test_path"}
 
-    client._upload_new(
-        file_path="test_file.csv",
+    client.upload_new(
+        file_path="tests/test_cases/countries.csv",
         mimetype="text/csv",
         chunk_size=5242880,
         alias_name="test_alias",
@@ -106,10 +96,10 @@ def test_upload_new(
     )
 
     mock_create_temp_chunks.assert_called_once_with(
-        Path("test_file.csv").absolute(), 5242880
+        Path("tests/test_cases/countries.csv").absolute(), 5242880
     )
     mock_generate_upload_new_params.assert_called_once_with(
-        Path("test_file.csv").absolute(),
+        Path("tests/test_cases/countries.csv").absolute(),
         "text/csv",
         5242880,
         "test_alias",
